requirements.txt
Flask==2.0.1
tensorflow==2.8.0
numpy==1.26.4
Pillow==9.5.0
werkzeug==2.0.3 